patch: maybe/ape-updates
APE updates required to support latest Python 2.7 build.

Notes:
The following files should be removed or verified that they do note exist:

/sys/include/ape/inttypes.h
/sys/src/ape/lib/bsd/sethostent.c
/sys/src/ape/lib/draw/libc.h
/sys/src/ape/lib/mp/port/libc.h
/sys/src/ape/lib/sec/port/9ape.c
/sys/src/ape/lib/sec/port/libc.h
/sys/src/ape/lib/sec/port/x509-ape.c
Sun Feb 23 01:36:08 EST 2014 geoff
    wow, a lot of changes.

Reference: /n/sources/patch/maybe/ape-updates
Date: Fri Aug 23 08:30:01 CES 2013

+++ sys/include/ape/stddef.h
@@ -17,7 +17,7 @@ typedef unsigned long size_t;
 #endif
 #ifndef _WCHAR_T
 #define _WCHAR_T
-typedef unsigned short wchar_t;
+typedef unsigned int wchar_t;
 #endif
 
 #endif /* __STDDEF_H */
